Make SSE the first choice for x86.
It is generally faster than 3DNowExt on AMD CPUs that support both (especially once Taihei brings in the improved dct64 for SSE).
The optimized dct36 seems even to make it _slower_ for K7 and K8. On K6 it is faster than plain dct36.
Need to check how it is on early K7 without SSE unit.